Refactoring, created separated package for EF Storage
