+++ src/DbLocalizationProvider.AspNetCore.EntityFramework/Entities/LocalizationResourceTranslationEntity.cs
@@ -4,7 +4,7 @@
 using System;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace DbLocalizationProvider.AspNetCore.Storage.Entities
+namespace DbLocalizationProvider.AspNetCore.EntityFramework.Entities
 {
     public class LocalizationResourceTranslationEntity
     {
