+++ src/DbLocalizationProvider.AspNetCore.EntityFramework/Extensions/DbContextOptionsBuilderExtensions.cs
@@ -1,11 +1,11 @@
 // Copyright (c) Valdis Iljuconoks, Andriy S'omak. All rights reserved.
 // Licensed under Apache-2.0. See the LICENSE file in the project root for more information
 
-using DbLocalizationProvider.AspNetCore.Storage.Customizers;
+using DbLocalizationProvider.AspNetCore.EntityFramework.Customizers;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 
-namespace DbLocalizationProvider.AspNetCore.Storage.Extensions
+namespace DbLocalizationProvider.AspNetCore.EntityFramework.Extensions
 {
     public static class DbContextOptionsBuilderExtensions
     {
