+++ src/DbLocalizationProvider.AspNetCore.EntityFramework/Handlers/AvailableLanguagesHandler.cs
@@ -4,11 +4,11 @@
 using System.Collections.Generic;
 using System.Globalization;
 using DbLocalizationProvider.Abstractions;
-using DbLocalizationProvider.AspNetCore.Storage.Repositories;
+using DbLocalizationProvider.AspNetCore.EntityFramework.Repositories;
 using DbLocalizationProvider.Cache;
 using DbLocalizationProvider.Queries;
 
-namespace DbLocalizationProvider.AspNetCore.Storage.Handlers
+namespace DbLocalizationProvider.AspNetCore.EntityFramework.Handlers
 {
     /// <summary>
     /// Reads all available languages form database (in which translations are added).
