+++ src/DbLocalizationProvider.AspNetCore.EntityFramework/Handlers/CreateOrUpdateTranslationHandler.cs
@@ -3,11 +3,11 @@
 
 using System;
 using DbLocalizationProvider.Abstractions;
-using DbLocalizationProvider.AspNetCore.Storage.Repositories;
+using DbLocalizationProvider.AspNetCore.EntityFramework.Repositories;
 using DbLocalizationProvider.Cache;
 using DbLocalizationProvider.Commands;
 
-namespace DbLocalizationProvider.AspNetCore.Storage.Handlers
+namespace DbLocalizationProvider.AspNetCore.EntityFramework.Handlers
 {
     /// <summary>
     /// Implementation for creating or updating existing translation
