+++ src/DbLocalizationProvider.AspNetCore.EntityFramework/Handlers/GetAllResourcesHandler.cs
@@ -1,16 +1,17 @@
 // Copyright (c) Valdis Iljuconoks, Andriy S'omak. All rights reserved.
 // Licensed under Apache-2.0. See the LICENSE file in the project root for more information
 
+using System.Collections.Generic;
 using DbLocalizationProvider.Abstractions;
-using DbLocalizationProvider.AspNetCore.Storage.Repositories;
+using DbLocalizationProvider.AspNetCore.EntityFramework.Repositories;
 using DbLocalizationProvider.Queries;
 
-namespace DbLocalizationProvider.AspNetCore.Storage.Handlers
+namespace DbLocalizationProvider.AspNetCore.EntityFramework.Handlers
 {
     /// <summary>
-    /// Retrieves single resource
+    /// Reads all resources from underlying storage
     /// </summary>
-    public class GetResourceHandler : IQueryHandler<GetResource.Query, LocalizationResource>
+    public class GetAllResourcesHandler : IQueryHandler<GetAllResources.Query, IEnumerable<LocalizationResource>>
     {
         /// <summary>
         /// Place where query handling happens
@@ -20,10 +21,10 @@ namespace DbLocalizationProvider.AspNetCore.Storage.Handlers
         /// You have to return something from the query execution. Of course you can return <c>null</c> as well if you
         /// will.
         /// </returns>
-        public LocalizationResource Execute(GetResource.Query query)
+        public IEnumerable<LocalizationResource> Execute(GetAllResources.Query query)
         {
             var repository = new ResourceRepository();
-            return repository.GetByKey(query.ResourceKey);
+            return repository.GetAll();
         }
     }
 }
