+++ src/DbLocalizationProvider.AspNetCore.EntityFramework/Handlers/GetResourceHandler.cs
@@ -1,17 +1,16 @@
 // Copyright (c) Valdis Iljuconoks, Andriy S'omak. All rights reserved.
 // Licensed under Apache-2.0. See the LICENSE file in the project root for more information
 
-using System.Collections.Generic;
 using DbLocalizationProvider.Abstractions;
-using DbLocalizationProvider.AspNetCore.Storage.Repositories;
+using DbLocalizationProvider.AspNetCore.EntityFramework.Repositories;
 using DbLocalizationProvider.Queries;
 
-namespace DbLocalizationProvider.AspNetCore.Storage.Handlers
+namespace DbLocalizationProvider.AspNetCore.EntityFramework.Handlers
 {
     /// <summary>
-    /// Reads all resources from underlying storage
+    /// Retrieves single resource
     /// </summary>
-    public class GetAllResourcesHandler : IQueryHandler<GetAllResources.Query, IEnumerable<LocalizationResource>>
+    public class GetResourceHandler : IQueryHandler<GetResource.Query, LocalizationResource>
     {
         /// <summary>
         /// Place where query handling happens
@@ -21,10 +20,10 @@ namespace DbLocalizationProvider.AspNetCore.Storage.Handlers
         /// You have to return something from the query execution. Of course you can return <c>null</c> as well if you
         /// will.
         /// </returns>
-        public IEnumerable<LocalizationResource> Execute(GetAllResources.Query query)
+        public LocalizationResource Execute(GetResource.Query query)
         {
             var repository = new ResourceRepository();
-            return repository.GetAll();
+            return repository.GetByKey(query.ResourceKey);
         }
     }
 }
