+++ src/DbLocalizationProvider.AspNetCore/ServiceLocator/HttpContextServiceProviderProxy.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace DbLocalizationProvider.AspNetCore.Storage.Locators
+namespace DbLocalizationProvider.AspNetCore.ServiceLocator
 {
     public class HttpContextServiceProviderProxy : IServiceProviderProxy
     {
