+++ tests/DbLocalizationProvider.Core.AspNetSample/Startup.cs
@@ -5,13 +5,12 @@ using DbLocalizationProvider.AdminUI.AspNetCore.Routing;
 using DbLocalizationProvider.AspNetCore;
 using DbLocalizationProvider.AspNetCore.ClientsideProvider;
 using DbLocalizationProvider.AspNetCore.ClientsideProvider.Routing;
+using DbLocalizationProvider.AspNetCore.EntityFramework.Extensions;
 using DbLocalizationProvider.AspNetCore.Extensions;
-using DbLocalizationProvider.AspNetCore.Storage.Extensions;
 using DbLocalizationProvider.Core.AspNet.ForeignAssembly;
 using DbLocalizationProvider.Core.AspNetSample.Data;
 using DbLocalizationProvider.Core.AspNetSample.Extensions;
 using DbLocalizationProvider.Core.AspNetSample.Resources;
-using DbLocalizationProvider.Storage.SqlServer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Identity;
